Adds battleeffects to set consistent timing
